+++ pusfume/scripts/mods/pusfume/_pusfume_native_config.lua
@@ -11,6 +11,7 @@ return {
     manual_clip_probe = false,
     manual_skin_probe = false,
     parent_child_material = false,
+    parent_child_package = false,
     root_animation_isolation = false,
     skin_name = "pusfume_skin",
     third_person_unit = "units/pusfume/pusfume_3p",